requirements.txt
google-cloud-datastore
tweepy
opencv-contrib-python
opencv-python
numpy
requests
